+++ WindowsFormsApp1/ExportCADDataTool.csproj
@@ -127,6 +127,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ExportFile.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="SolidworksAPIControl.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
